backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pandas
scikit-learn
joblib
pydantic